Styling update, still working on decimal issue

+++ Operations.cs
@@ -8,24 +8,24 @@ namespace SimpleCalculator
 {
     public class Operations
     {
-        public decimal Add(int x, int y)
+        public float Add(int x, int y)
         {
-            decimal z = x + y;
+            float z = x + y;
             return z;
         }
-        public decimal Mult(int x, int y)
+        public float Mult(int x, int y)
         {
-            decimal z = x * y;
+            float z = x * y;
             return z;
         }
-        public decimal Div(int x, int y)
+        public float Div(int x, int y)
         {
-            decimal z = x / y;
+            float z = x / y;
             return z;
         }
-        public decimal Sub(int x, int y)
+        public float Sub(int x, int y)
         {
-            decimal z = x - y;
+            float z = x - y;
             return z;
         }
     }
